// src/interfaces/parameter.ts
export interface IParameters {
  menopausal_status: number | null | null;
  platelets: number | null;
  albumin: number | null;
  globulin: number | null;
  solid: number | null;
  papillary_projection: number | null;
  cyst_wall: number | null;
  septum: number | null;
  // ascites: number | null;
  ca_125: number | null;
}